requirements.txt
psutil
py-cord
python-dotenv
py-cpuinfo